enhance post controller: implement user authentication and authorization; refine error handling across methods

+++ src/controllers/post.controller.js
@@ -3,8 +3,14 @@ import postService from "../services/post.service.js";
 class PostController {
     async createPost(req, res, next) {
         try {
-            const post = await postService.createPost(req.params.author, req.body);
-            return res.status(201).json(post);
+            const login = req.principal.username;
+            const author = req.params.author
+            const isSelf = login === author;
+            if (!isSelf) {
+                return next({message: 'Invalid credentials', statusCode: 403});
+            }
+                const post = await postService.createPost(req.params.author, req.body);
+                return res.status(201).json(post);
         } catch (error) {
             return next(error);
         }
@@ -12,32 +18,46 @@ class PostController {
 
     async getPostById(req, res, next) {
         try {
-            const post = await postService.getPostById(req.params.id);
-            return res.json(post);
+            const login = req.principal.username;
+            if(login) {
+                const post = await postService.getPostById(req.params.id);
+                return res.json(post);
+            }
         } catch (error) {
             return next(error);
         }
     }
 
-    async deletePost(req, res, next) {
+    async deletePost(req, res, next) { //done
         try {
-            const post = await postService.deletePost(req.params.id);
-            return res.json(post);
+            const isModerator = req.principal.roles.includes('MODERATOR');
+            const principalLogin = req.principal.username;
+            const {author} = await postService.getPostById(req.params.id);
+            const isSelf = principalLogin === author;
+            if(!isModerator && !isSelf) {
+                return next({message: 'Invalid credentials', statusCode: 403});
+            }
+                const post = await postService.deletePost(req.params.id);
+                return res.json(post);
+
         } catch (error) {
             return next(error);
         }
     }
 
-    async addLike(req, res, next) {
+    async addLike(req, res, next) { //done
         try {
-            await postService.addLike(req.params.id);
-            return res.sendStatus(204);
+            const login = req.principal.username;
+            if(login) {
+                await postService.addLike(req.params.id);
+                return res.sendStatus(204);
+            }
         } catch (err) {
             return next(err);
         }
     }
 
-    async getPostsByAuthor(req, res, next) {
+    async getPostsByAuthor(req, res, next) { //done
         try {
             const posts = await postService.getPostsByAuthor(req.params.author);
             return res.json(posts);
@@ -48,6 +68,7 @@ class PostController {
 
     async addComment(req, res, next) {
         try {
+            //TODO
             const post = await postService.addComment(req.params.id, req.params.commenter, req.body.message);
             return res.json(post);
         } catch (err) {
@@ -55,7 +76,7 @@ class PostController {
         }
     }
 
-    async getPostsByTags(req, res, next) {
+    async getPostsByTags(req, res, next) { //done
         let values;
         if (Array.isArray(req.query.values)) {
             values = req.query.values.reduce((acc, item) => acc + ',' + item);
@@ -70,7 +91,7 @@ class PostController {
         }
     }
 
-    async getPostsByPeriod(req, res, next) {
+    async getPostsByPeriod(req, res, next) { //done
         try {
             const {dateFrom, dateTo} = req.query;
             const posts = await postService.getPostsByPeriod(dateFrom, dateTo);
@@ -82,6 +103,13 @@ class PostController {
 
     async updatePost(req, res, next) {
         try {
+            const login = req.principal.username;
+            const {author} = await postService.getPostById(req.params.id);
+            console.log(author);
+           const isSelf = login === author;
+            if (!isSelf) {
+                return next({message: 'Invalid credentials', statusCode: 403});
+            }
             const post = await postService.updatePost(req.params.id, req.body);
             return res.json(post);
         } catch (err) {
